drivers: modem: make modem cmd macros and structures public
Make modem cmd macros and structures public so that
applications can use them. That is also necessary
for the upcoming support of custom modem setup commands.
